+++ src/object_size.h
@@ -5,6 +5,6 @@
 
 #include <jvmti.h>
 
-jint estimateObjectSize(JNIEnv *jni, jvmtiEnv *jvmti, jclass thisClass, jobject object);
+jlong estimateObjectSize(JNIEnv *jni, jvmtiEnv *jvmti, jclass thisClass, jobject object);
 
 #endif //NATIVE_MEMORY_AGENT_OBJECT_SIZE_H
